business-ai-copilot/requirements.txt
streamlit
sqlalchemy
pymssql
python-dotenv
pypdf
pandas
google-generativeai
